+++ packages/presets/src/theme/recipes/tooltip.ts
@@ -8,12 +8,19 @@ export const tooltip = defineSlotRecipe({
     content: {
       background: 'fg.default',
       borderRadius: 'l2',
-      boxShadow: 'lg',
+      boxShadow: 'sm',
       color: 'bg.default',
       fontWeight: 'semibold',
       px: '3',
       py: '2',
       textStyle: 'xs',
+      maxWidth: '2xs',
+      _open: {
+        animation: 'fadeIn 0.25s ease-out',
+      },
+      _closed: {
+        animation: 'fadeOut 0.2s ease-out',
+      },
     },
   },
 })