+++ website/src/app/docs/layout.tsx
@@ -5,15 +5,16 @@ import { Sidebar } from '~/components/layout/sidebar'
 const DocsLayout = (props: PropsWithChildren) => {
   const { children } = props
   return (
-    <Container flex="1">
+    <Container flex="1" px="0">
       <Flex
         display={{ base: 'none', lg: 'flex' }}
         position="fixed"
-        width="60"
         top="74px"
         bottom="0"
+        minW="60"
         pt="10"
         overflowY="auto"
+        px={{ base: '4', md: '6' }}
       >
         <Sidebar />
       </Flex>